// apps/variant-browser/src/app/services/data-source/models/consequence-line.ts
export interface ConsequenceLine {
  impact?: string;
  enst?: string;
  dna_ncbi?: string;
  rna_ncbi?: string;
  start?: string;
  end?: string;
  sift?: string;
  effectName?: string;
  effectAccession?: string;
  effectDescription?: string;
  hgvsp?: string;
  hgvsc?: string;
  polyphen?: string;
  geneName?: string;
  ensg?: string;
  hgnc?: string;
  ncbi?: string;
  symbol?: string;
  biotypeAccession?: string;
  biotypeName?: string;
  biotypeDescription?: string;
}